+++ client/src/main/java/de/fraunhofer/iosb/client/policy/PolicyController.java
@@ -16,9 +16,9 @@
 package de.fraunhofer.iosb.client.policy;
 
 import de.fraunhofer.iosb.client.util.Pair;
-import org.eclipse.edc.catalog.spi.Dataset;
-import org.eclipse.edc.connector.policy.spi.PolicyDefinition;
-import org.eclipse.edc.connector.spi.catalog.CatalogService;
+import org.eclipse.edc.connector.controlplane.catalog.spi.Dataset;
+import org.eclipse.edc.connector.controlplane.policy.spi.PolicyDefinition;
+import org.eclipse.edc.connector.controlplane.services.spi.catalog.CatalogService;
 import org.eclipse.edc.policy.model.Policy;
 import org.eclipse.edc.spi.monitor.Monitor;
 import org.eclipse.edc.spi.system.configuration.Config;