+++ edc-extension4aas/src/main/java/de/fraunhofer/iosb/app/dataplane/aas/pipeline/AasDataSourceFactory.java
@@ -1,0 +1,75 @@
+/*
+ * Copyright (c) 2021 Fraunhofer IOSB, eine rechtlich nicht selbstaendige
+ * Einrichtung der Fraunhofer-Gesellschaft zur Foerderung der angewandten
+ * Forschung e.V.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package de.fraunhofer.iosb.app.dataplane.aas.pipeline;
+
+import de.fraunhofer.iosb.app.util.HttpRestClient;
+import org.eclipse.edc.connector.dataplane.http.params.HttpRequestFactory;
+import org.eclipse.edc.connector.dataplane.http.spi.HttpDataAddress;
+import org.eclipse.edc.connector.dataplane.http.spi.HttpRequestParamsProvider;
+import org.eclipse.edc.connector.dataplane.spi.pipeline.DataSource;
+import org.eclipse.edc.connector.dataplane.spi.pipeline.DataSourceFactory;
+import org.eclipse.edc.spi.monitor.Monitor;
+import org.eclipse.edc.spi.result.Result;
+import org.eclipse.edc.spi.types.domain.transfer.DataFlowStartMessage;
+import org.jetbrains.annotations.NotNull;
+
+/**
+ * Inspired by {@link org.eclipse.edc.connector.dataplane.http.pipeline.HttpDataSourceFactory}
+ */
+public class AasDataSourceFactory implements DataSourceFactory {
+
+    public static final String AAS_DATA_TYPE = "AasData";
+
+    private final Monitor monitor;
+    private final HttpRequestParamsProvider requestParamsProvider;
+    private final HttpRequestFactory requestFactory;
+
+    public AasDataSourceFactory(HttpRequestParamsProvider requestParamsProvider, Monitor monitor) {
+        this.monitor = monitor;
+        this.requestParamsProvider = requestParamsProvider;
+        this.requestFactory = new HttpRequestFactory();
+    }
+
+    @Override
+    public boolean canHandle(DataFlowStartMessage request) {
+        return AAS_DATA_TYPE.equals(request.getSourceDataAddress().getType());
+    }
+
+    @Override
+    public DataSource createSource(DataFlowStartMessage request) {
+        var dataAddress = HttpDataAddress.Builder.newInstance()
+                .copyFrom(request.getSourceDataAddress())
+                .build();
+        return AasDataSource.Builder.newInstance()
+                .httpClient(HttpRestClient.getInstance())
+                .monitor(monitor)
+                .requestId(request.getId())
+                .name(dataAddress.getName())
+                .params(requestParamsProvider.provideSourceParams(request))
+                .requestFactory(requestFactory)
+                .build();
+    }
+
+    @Override
+    public @NotNull Result<Void> validateRequest(DataFlowStartMessage request) {
+        try {
+            createSource(request);
+        } catch (Exception e) {
+            return Result.failure("Failed to build HttpDataSource: " + e.getMessage());
+        }
+        return Result.success();
+    }
+}