+++ changelog.md
@@ -2,7 +2,9 @@
 
 ## Current development version
 
-Compatibility: **Eclipse Dataspace Connector v0.5.1**
+Compatibility: **Eclipse Dataspace Connector v0.6.2**
+
+Compatibility: **AAS V3**
 
 **New Features**
 
@@ -10,6 +12,12 @@ Compatibility: **Eclipse Dataspace Connector v0.5.1**
 
 
 **Bugfixes**
+
+- DataTransfer with self-signed certificates (FA³ST DataSource)
+
+**Miscellaneous**
+
+- !Change API port for self-description to 8281/9291 in examples!
 
 ## V1.0.0-alpha5
 
@@ -109,7 +117,7 @@ This version is compatible to **Eclipse Dataspace Connector v0.0.1-Milestone 7**
 **Bugfixes**
 
 * CustomAuthenticationRequestFilter
-    * No requests were filtered when self description was exposed by config
+    * No requests were filtered when self-description was exposed by config
 
 ## V1.0.0-alpha
 
@@ -126,7 +134,7 @@ This version is compatible to **Eclipse Dataspace Connector v0.0.1-Milestone 6**
       AuthenticationServices and AuthenticationRequestFilters.
 
 * Configuration value "exposeSelfDescription"
-    * Set to _False_ to not expose the extension's self descriptions
+    * Set to _False_ to not expose the extension's self-descriptions
     * Default value is _True_
 
 **Bugfixes**